Separated components and styles in different folders.
